feat: use env for version

+++ pack.sh
@@ -8,26 +8,35 @@ if [[ $# -eq 0 ]] ; then
   exit 1
 fi
 
+ver="${GOVERSION:-1.20.2}"
+archive="go${ver}.linux-riscv64.tar.gz"
 
-VERSION=$1
-
-pushd "./out/${VERSION}/go/bin/"
-  bindir=linux_riscv64
-  if [ -d "${bindir}" ]
+hash () {
+    sha256sum "${archive}" > "${archive}.sha256.txt"
+    cat "${archive}.sha256.txt"
+    ls --human-readable --kibibytes -Sl "${archive}"
+}
+pushd out
+  archive="go${ver}.linux-riscv64.tar.gz"
+  if [ -f "${archive}" ]
   then
-    echo "Moving binaries from ${bindir}"
-    mv ${bindir}/* .
-    rm -rf ${bindir}
+    echo "Already built as ${archive}"
+    hash
+    exit 0
   fi
-  file go gofmt
-popd
 
-pushd ./out
-  archive="go${VERSION}.linux-riscv64.tar.gz"
-  rm -f "${archive}"
+  pushd "${ver}/go/bin/"
+    bindir=linux_riscv64
+    if [ -d "${bindir}" ]
+    then
+      echo "Moving binaries from ${bindir}"
+      mv ${bindir}/* .
+      rm -rf ${bindir}
+    fi
+    file go gofmt
+  popd
+
   echo "Creating ${archive}"
-  tar -C "${VERSION}" -czf "${archive}" go
-  sha256sum "${archive}" > "${archive}.sha256.txt"
-  cat "${archive}.sha256.txt"
-  ls --human-readable --kibibytes -Sl "${archive}"
+  tar -C "${ver}" -czf "${archive}" go
+  hash
 popd
